requirements.txt
Django==3.2.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
whitenoise==6.2.0
